requirements.txt
Flask==3.0.2
cassandra-driver==3.29.1
python-dotenv==1.0.1
pydantic==2.6.3
rq
redis
fastapi==0.110.1
uvicorn[standard]==0.29.0
datetime


